Adjust i18n tests
